Add `buttonSize` prop
