+++ fiberplane-ui/src/components/Button/types.ts
@@ -1,0 +1,8 @@
+export type ButtonSize = "small" | "medium";
+
+export type ButtonStyle =
+  | "primary"
+  | "secondary"
+  | "tertiary-color"
+  | "tertiary-grey"
+  | "danger";
