+++ fiberplane-ui/src/components/Select/Select.tsx
@@ -6,7 +6,7 @@ import { css, styled } from "styled-components";
 import { useOpen } from "@fiberplane/hooks";
 import { cancelEvent, noop } from "../../utils";
 import { AttachedPopup } from "../AttachedPopup";
-import { Button } from "../Button";
+import { Button } from "../Button/Button";
 import { Icon } from "../Icon";
 import { SelectOption as Option } from "./SelectOption";
