+++ fiberplane-ui/src/components/index.ts
@@ -1,5 +1,5 @@
 export * from "./AttachedPopup";
-export * from "./Button";
+export * from "./Button/Button";
 export * from "./Icon";
 export * from "./IconButton";
 export * from "./Input";
